Move view, template and validator to 'core' app + add errors handling
